Create Number Game
This is a game where a random number is generated between 0 to 100 and the user have to guess that number. User have 3 attempts to guess the correct number. If user will not be able to guess the correct number then they will have the option to either play again or to leave.